+++ packages/components/Checkbox/src/Checkbox.tsx
@@ -31,13 +31,13 @@ export const Checkbox = compose<ICheckboxType>({
     const buttonRef = useViewCommandFocus(userProps.componentRef);
 
     // Handles the "Space" key toggling the Checkbox
-    const onKeyUpSpace = useKeyCallback(' ', toggleChecked);
+    const onKeyUpSpace = useKeyCallback(toggleChecked, ' ');
 
     const state: ICheckboxState = {
       ...pressable.state,
       disabled,
       checked: isChecked,
-      boxAtEnd: boxSide == undefined || boxSide == 'start' ? false : true
+      boxAtEnd: boxSide == undefined || boxSide == 'start' ? false : true,
     };
 
     // Grab the styling information from the userProps, referencing the state as well as the props.
@@ -52,7 +52,7 @@ export const Checkbox = compose<ICheckboxType>({
             break;
         }
       },
-      [toggleChecked, userProps, state, pressable.props]
+      [toggleChecked, userProps, state, pressable.props],
     );
 
     const slotProps = mergeSettings<ICheckboxSlotProps>(styleProps, {
@@ -62,14 +62,14 @@ export const Checkbox = compose<ICheckboxType>({
         ...pressable.props,
         accessibilityRole: 'checkbox',
         accessibilityLabel: ariaLabel || label,
-        accessibilityState: {disabled: state.disabled, checked: state.checked },
+        accessibilityState: { disabled: state.disabled, checked: state.checked },
         accessibilityActions: [{ name: 'Toggle', label: checkboxSelectActionLabel }],
         onAccessibilityAction: onAccessibilityAction,
-        onKeyUp: onKeyUpSpace
+        onKeyUp: onKeyUpSpace,
       },
       // Temporary checkmark until SVG functionality
       checkmark: { children: '✓' },
-      content: { children: label }
+      content: { children: label },
     });
 
     return { slotProps, state };
@@ -93,7 +93,7 @@ export const Checkbox = compose<ICheckboxType>({
     root: View,
     checkbox: { slotType: View, filter: filterViewProps },
     checkmark: Text,
-    content: Text
+    content: Text,
   },
   styles: {
     root: [],
@@ -102,24 +102,18 @@ export const Checkbox = compose<ICheckboxType>({
       borderTokens,
       [
         { source: 'checkboxBackgroundColor', lookup: getPaletteFromTheme, target: 'backgroundColor' },
-        { source: 'checkboxBorderColor', lookup: getPaletteFromTheme, target: 'borderColor' }
-      ]
+        { source: 'checkboxBorderColor', lookup: getPaletteFromTheme, target: 'borderColor' },
+      ],
     ],
     checkmark: [
       foregroundColorTokens,
       [
         { source: 'checkmarkColor', lookup: getPaletteFromTheme, target: 'color' },
-        { source: 'checkmarkVisibility', target: 'opacity' }
-      ]
+        { source: 'checkmarkVisibility', target: 'opacity' },
+      ],
     ],
-    content: [
-      foregroundColorTokens,
-      textTokens,
-      [
-        { source: 'textBorderColor', lookup: getPaletteFromTheme, target: 'borderColor' }
-      ]
-    ]
-  }
+    content: [foregroundColorTokens, textTokens, [{ source: 'textBorderColor', lookup: getPaletteFromTheme, target: 'borderColor' }]],
+  },
 });
 
 export default Checkbox;