+++ packages/experimental/Button/src/CompoundButton/CompoundButton.test.tsx
@@ -1,0 +1,8 @@
+import * as React from 'react';
+import { CompoundButton } from './CompoundButton';
+import * as renderer from 'react-test-renderer';
+
+it('CompoundButton default', () => {
+  const tree = renderer.create(<CompoundButton content="Default Button" secondaryContent="sublabel" />).toJSON();
+  expect(tree).toMatchSnapshot();
+});