+++ packages/theming/theme-types/src/Theme.types.ts
@@ -1,6 +1,10 @@
-import { ThemeColorDefinition } from './Color.types';
+import { ColorValue, ThemeColorDefinition } from './Color.types';
 import { OfficePalette } from './palette.types';
 import { Typography, PartialTypography } from './Typography.types';
+
+type TwoLevelPartial<T> = {
+  [K in keyof T]?: Partial<T[K]>;
+};
 
 export interface Spacing {
   s2: string;
@@ -18,25 +22,22 @@ export interface Theme {
   colors: ThemeColorDefinition;
   typography: Typography;
   components: { [key: string]: object };
-  componentTokens?: object;
   spacing: Spacing;
   host: {
+    // appearance of the theme, this corresponds to the react-native Appearance library values, though can be overwritten
+    appearance: 'light' | 'dark';
+
     // Office palette, if running in Office with the native module connected in the theme
     palette?: OfficePalette;
+    colors?: { [key: string]: ColorValue };
   };
 }
 
 /**
- * A partially specified theme.
- *
- * Useful for overriding specific visual elements in a fully specified theme.
+ * Generally a partial theme is comprised of partial versions of the objects within the theme, with the exception of typography
+ * which has additional levels of hierarchy
  */
-export interface PartialTheme {
-  name?: string;
-  colors?: Partial<ThemeColorDefinition>;
+export type PartialTheme = Omit<TwoLevelPartial<Theme>, 'typography' | 'host'> & {
   typography?: PartialTypography;
-  components?: { [key: string]: object };
-  componentTokens?: object;
-  spacing?: Spacing;
-  host?: object;
-}
+  host?: TwoLevelPartial<Theme['host']>;
+};