+++ packages/components/Checkbox/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @fluentui-react-native/checkbox
 
-This log was last generated on Fri, 25 Sep 2020 19:21:43 GMT and should not be manually modified.
+This log was last generated on Mon, 19 Oct 2020 22:06:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.7.1
+
+Mon, 19 Oct 2020 22:06:02 GMT
+
+### Patches
+
+- rework theme usage in tester (jasonmo@microsoft.com)
+
+## 0.7.0
+
+Tue, 13 Oct 2020 01:39:45 GMT
+
+### Minor changes
+
+- remove useKeyCallback import (lehon@microsoft.com)
+
+## 0.6.0
+
+Mon, 12 Oct 2020 04:09:31 GMT
+
+### Minor changes
+
+- shared useKeyCallback (email not defined)
+
+## 0.5.10
+
+Mon, 28 Sep 2020 20:32:21 GMT
+
+### Patches
+
+- move test pages (aliciakds88@gmail.com)
 
 ## 0.5.8