+++ packages/components/RadioGroup/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/radio-group
 
-This log was last generated on Fri, 25 Sep 2020 19:21:43 GMT and should not be manually modified.
+This log was last generated on Mon, 19 Oct 2020 22:06:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.5.1
+
+Mon, 19 Oct 2020 22:06:02 GMT
+
+### Patches
+
+- rework theme usage in tester (jasonmo@microsoft.com)
+
+## 0.5.0
+
+Mon, 12 Oct 2020 04:09:31 GMT
+
+### Minor changes
+
+- shared useKeyCallback (email not defined)
+
+## 0.4.15
+
+Wed, 07 Oct 2020 17:29:29 GMT
+
+### Patches
+
+- Adding FocusZone to RadioGroup (safreibe@microsoft.com)
 
 ## 0.4.10