+++ packages/components/Separator/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/separator
 
-This log was last generated on Fri, 25 Sep 2020 19:21:43 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Oct 2020 04:09:31 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.0
+
+Mon, 12 Oct 2020 04:09:31 GMT
+
+### Minor changes
+
+- shared useKeyCallback (email not defined)
+
+## 0.5.0
+
+Thu, 01 Oct 2020 16:54:19 GMT
+
+### Minor changes
+
+- switch Separator to new framework (jasonmo@microsoft.com)
 
 ## 0.4.12