+++ packages/components/text/CHANGELOG.md
@@ -1,8 +1,24 @@
 # Change Log - @fluentui-react-native/text
 
-This log was last generated on Fri, 25 Sep 2020 19:21:43 GMT and should not be manually modified.
+This log was last generated on Mon, 19 Oct 2020 22:06:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.6.19
+
+Mon, 19 Oct 2020 22:06:02 GMT
+
+### Patches
+
+- rework theme usage in tester (jasonmo@microsoft.com)
+
+## 0.6.18
+
+Mon, 12 Oct 2020 04:09:31 GMT
+
+### Patches
+
+- shared hooks (email not defined)
 
 ## 0.6.12