+++ packages/theming/default-theme/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @fluentui-react-native/default-theme
 
-This log was last generated on Wed, 23 Sep 2020 18:31:48 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Oct 2020 04:09:31 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.0
+
+Mon, 12 Oct 2020 04:09:31 GMT
+
+### Minor changes
+
+- button semantic colors (email not defined)
+
+## 0.2.0
+
+Mon, 28 Sep 2020 20:32:21 GMT
+
+### Minor changes
+
+- add more host colors (aliciakds88@gmail.com)
+
+## 0.1.5
+
+Sun, 27 Sep 2020 04:08:27 GMT
+
+### Patches
+
+- add light and dark awareness to default themes (jasonmo@microsoft.com)
 
 ## 0.1.3