+++ packages/theming/theme-types/CHANGELOG.md
@@ -1,8 +1,48 @@
 # Change Log - @fluentui-react-native/theme-types
 
-This log was last generated on Fri, 25 Sep 2020 18:01:57 GMT and should not be manually modified.
+This log was last generated on Mon, 19 Oct 2020 22:06:02 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.3.1
+
+Mon, 19 Oct 2020 22:06:02 GMT
+
+### Patches
+
+- rework theme usage in tester (jasonmo@microsoft.com)
+
+## 0.3.0
+
+Mon, 12 Oct 2020 04:09:31 GMT
+
+### Minor changes
+
+- button semantic colors (email not defined)
+
+## 0.2.1
+
+Tue, 29 Sep 2020 18:14:46 GMT
+
+### Patches
+
+- remove extra componentTokens reference and route everything to components (jasonmo@microsoft.com)
+
+## 0.2.0
+
+Mon, 28 Sep 2020 20:32:21 GMT
+
+### Minor changes
+
+- add more host colors (aliciakds88@gmail.com)
+
+## 0.1.5
+
+Sun, 27 Sep 2020 04:08:27 GMT
+
+### Patches
+
+- add light and dark awareness to default themes (jasonmo@microsoft.com)
 
 ## 0.1.4