+++ packages/theming/win32-theme/CHANGELOG.md
@@ -1,8 +1,40 @@
 # Change Log - @fluentui-react-native/win32-theme
 
-This log was last generated on Fri, 25 Sep 2020 18:01:57 GMT and should not be manually modified.
+This log was last generated on Mon, 12 Oct 2020 04:09:31 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 0.4.0
+
+Mon, 12 Oct 2020 04:09:31 GMT
+
+### Minor changes
+
+- button semantic colors (email not defined)
+
+## 0.3.2
+
+Wed, 07 Oct 2020 18:52:49 GMT
+
+### Patches
+
+- Fixing redbox on macOS from theming module (nakambo@microsoft.com)
+
+## 0.3.0
+
+Mon, 28 Sep 2020 20:32:21 GMT
+
+### Minor changes
+
+- add more host colors (aliciakds88@gmail.com)
+
+## 0.2.0
+
+Sun, 27 Sep 2020 04:08:27 GMT
+
+### Minor changes
+
+- add light and dark awareness to default themes (jasonmo@microsoft.com)
 
 ## 0.1.2